+++ integration/spark/app/src/main/java/io/openlineage/spark/agent/lifecycle/SparkSQLExecutionContext.java
@@ -53,7 +53,9 @@ class SparkSQLExecutionContext implements ExecutionContext {
   private final OpenLineageContext olContext;
   private final EventEmitter eventEmitter;
   private final OpenLineageRunEventBuilder runEventBuilder;
+  private final boolean isInStreamingQueryMode;
 
+  private boolean shouldEmitStartEvent = true;
   private boolean emittedOnSqlExecutionStart = false;
   private boolean emittedOnSqlExecutionEnd = false;
   private boolean emittedOnJobStart = false;
@@ -67,11 +69,13 @@ class SparkSQLExecutionContext implements ExecutionContext {
       long executionId,
       EventEmitter eventEmitter,
       OpenLineageContext olContext,
-      OpenLineageRunEventBuilder runEventBuilder) {
+      OpenLineageRunEventBuilder runEventBuilder,
+      boolean isInStreamingQueryMode) {
     this.executionId = executionId;
     this.eventEmitter = eventEmitter;
     this.olContext = olContext;
     this.runEventBuilder = runEventBuilder;
+    this.isInStreamingQueryMode = isInStreamingQueryMode;
   }
 
   @Override
@@ -90,7 +94,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
 
     olContext.setActiveJobId(activeJobId);
     // only one START event is expected, in case it was already sent with jobStart, we send running
-    EventType eventType = emittedOnJobStart ? RUNNING : START;
+    EventType eventType = !shouldEmitStartEvent || emittedOnJobStart ? RUNNING : START;
     emittedOnSqlExecutionStart = true;
 
     RunEvent event =
@@ -134,6 +138,9 @@ class SparkSQLExecutionContext implements ExecutionContext {
     EventType eventType;
     if (emittedOnJobStart && !emittedOnJobEnd) {
       // expecting jobEnd event later on
+      eventType = RUNNING;
+    } else if (isInStreamingQueryMode) {
+      // streaming query mode should never emit COMPLETE event for SQL
       eventType = RUNNING;
     } else {
       eventType = COMPLETE;
@@ -253,7 +260,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
 
     // only one START event is expected, in case it was already sent with sqlExecutionStart, we send
     // running
-    EventType eventType = emittedOnSqlExecutionStart ? RUNNING : START;
+    EventType eventType = !shouldEmitStartEvent || emittedOnSqlExecutionStart ? RUNNING : START;
     emittedOnJobStart = true;
 
     RunEvent event =
@@ -300,6 +307,9 @@ class SparkSQLExecutionContext implements ExecutionContext {
     } else if (emittedOnSqlExecutionStart && !emittedOnSqlExecutionEnd) {
       // still waiting for sqlExecutionEnd event which will emit COMPLETE event
       eventType = RUNNING;
+    } else if (isInStreamingQueryMode) {
+      // streaming query mode should never emit COMPLETE event for SQL
+      eventType = RUNNING;
     } else {
       eventType = COMPLETE;
     }
@@ -335,6 +345,10 @@ class SparkSQLExecutionContext implements ExecutionContext {
         eventEmitter.getApplicationRunId(),
         eventEmitter.getApplicationJobName(),
         eventEmitter.getJobNamespace());
+  }
+
+  public void setStartEmitted() {
+    this.shouldEmitStartEvent = false;
   }
 
   protected OpenLineage.JobBuilder buildJob() {