+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/SparkStreamingTest.java
@@ -16,6 +16,7 @@ import static org.apache.spark.sql.functions.expr;
 import static org.apache.spark.sql.functions.from_json;
 import static org.apache.spark.sql.functions.from_unixtime;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -33,6 +34,7 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
 import java.util.Random;
@@ -62,13 +64,13 @@ import org.apache.spark.sql.streaming.StreamingQuery;
 import org.apache.spark.sql.streaming.StreamingQueryException;
 import org.apache.spark.sql.streaming.Trigger;
 import org.apache.spark.sql.types.StructType;
-import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.testcontainers.containers.KafkaContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.shaded.com.google.common.collect.ImmutableMap;
+import org.testcontainers.shaded.org.apache.commons.lang3.tuple.Pair;
 import org.testcontainers.utility.DockerImageName;
 
 @Slf4j
@@ -176,14 +178,16 @@ class SparkStreamingTest {
 
     kafkaContainer.close();
 
-    Awaitility.await().atMost(Duration.ofSeconds(60)).until(() -> !kafkaContainer.isRunning());
-    Awaitility.await().atMost(Duration.ofSeconds(60)).until(() -> spark.sparkContext().isStopped());
+    await().atMost(Duration.ofSeconds(60)).until(() -> !kafkaContainer.isRunning());
+    await().atMost(Duration.ofSeconds(60)).until(() -> spark.sparkContext().isStopped());
 
     List<RunEvent> events =
         handler.getEventsMap().getOrDefault("test_kafka_source_to_kafka_sink", new ArrayList<>());
 
     List<RunEvent> sqlEvents =
         events.stream()
+            .filter(
+                x -> x.getJob().getFacets() != null && x.getJob().getFacets().getJobType() != null)
             .filter(
                 x -> "STREAMING".equals(x.getJob().getFacets().getJobType().getProcessingType()))
             .collect(Collectors.toList());
@@ -236,6 +240,7 @@ class SparkStreamingTest {
 
   @Test
   @EnabledIfSystemProperty(named = SPARK_VERSION, matches = SPARK_3_OR_ABOVE)
+  @SuppressWarnings("PMD.JUnitTestContainsTooManyAsserts")
   void testKafkaSourceToBatchSink() throws TimeoutException, StreamingQueryException, IOException {
     KafkaTestContainer kafkaContainer = setupKafkaContainer();
 
@@ -268,12 +273,17 @@ class SparkStreamingTest {
             .start();
 
     streamingQuery.awaitTermination(Duration.ofSeconds(20).toMillis());
+    streamingQuery.stop(); // triggers QueryTerminatedEvent event
+
+    waitForCompleteEvent("test_kafka_source_to_batch_sink");
     List<RunEvent> events = handler.getEventsMap().get("test_kafka_source_to_batch_sink");
 
     assertThat(events).isNotEmpty();
 
     List<RunEvent> kafkaInputEvents =
-        events.stream().filter(x -> !x.getInputs().isEmpty()).collect(Collectors.toList());
+        events.stream()
+            .filter(x -> x.getInputs() != null && !x.getInputs().isEmpty())
+            .collect(Collectors.toList());
 
     assertThat(kafkaInputEvents).isNotEmpty();
 
@@ -285,7 +295,9 @@ class SparkStreamingTest {
         });
 
     List<RunEvent> outputEvents =
-        events.stream().filter(x -> !x.getOutputs().isEmpty()).collect(Collectors.toList());
+        events.stream()
+            .filter(x -> x.getOutputs() != null && !x.getOutputs().isEmpty())
+            .collect(Collectors.toList());
 
     assertThat(outputEvents).isNotEmpty();
 
@@ -296,8 +308,44 @@ class SparkStreamingTest {
           assertEquals("file", event.getOutputs().get(0).getNamespace());
         });
 
+    // make sure events share the same runUuid and jobName
+    assertThat(
+            events.stream()
+                .filter(
+                    e ->
+                        e.getRun().getFacets() == null
+                            || e.getRun().getFacets().getParent()
+                                != null) // ignore application events
+                .map(x -> Pair.of(x.getRun().getRunId(), x.getJob().getName()))
+                .distinct()
+                .collect(Collectors.toList()))
+        .hasSize(1);
+
+    List<String> eventTypes =
+        events.stream()
+            .filter(
+                e ->
+                    e.getRun().getFacets() == null
+                        || e.getRun().getFacets().getParent() != null) // ignore application events
+            .map(e -> e.getEventType().toString())
+            .collect(Collectors.toList());
+
+    // there should be ONLY one START event and one COMPLETE event
+    assertThat(eventTypes).containsOnlyOnceElementsOf(Arrays.asList("START", "COMPLETE"));
+
     kafkaContainer.stop();
     spark.stop();
+  }
+
+  private static void waitForCompleteEvent(String jobName) {
+    await()
+        .pollInterval(Duration.ofMillis(500))
+        .atMost(Duration.ofSeconds(10))
+        .until(
+            () ->
+                handler.getEventsMap().getOrDefault(jobName, Collections.emptyList()).stream()
+                    .map(r -> r.getEventType().toString())
+                    .anyMatch("COMPLETE"::equalsIgnoreCase));
   }
 
   @Test